requirements.txt
pkce==1.0.3
python-dotenv==1.2.1
pyyaml==6.0.3
requests==2.32.5